backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
groq==0.15.0
httpcore==1.0.5
httpx==0.28.1
jsonschema==4.22.0
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-groq==0.2.3
langchain-text-splitters==0.3.5
langsmith==0.2.10
pydantic==2.10.5
pyparsing==3.1.1
PyPDF2==3.0.1
python-dotenv==1.0.1
python-json-logger==2.0.7
regex==2025.11.6
requests==2.32.3
requests-toolbelt==1.0.0
toml==0.10.2
typer==0.16.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
pymongo==4.4.1
flask==3.1.0
gunicorn==20.1.0
pdfplumber==0.9.0
PyMuPDF>=1.24.0
celery==5.3.6
google-api-python-client==2.85.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
cloudinary==1.41.0
aiosmtplib==1.1.6
sib-api-v3-sdk==7.6.0
pyotp==2.9.0
Flask-Mail==0.9.1
bcrypt==4.1.2
